extraer resolvers a su propio archivo
